requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
gunicorn==21.2.0
prometheus-flask-exporter==0.23.0
requests==2.31.0